refactor(KEN-1032): a check and a write never run together
The commit counter that let a check's stale report lose to a commit is
gone. Check refuses while a write is out and each write refuses while a
check is out, so no report is ever built before a commit it misses.

updates-writes.ts stays: it gained the removal account (saying) in
KEN-709 after the audit that filed this, and eight write paths ride on
it. Only the counter left.
