+++ changelog.d/fixed/KEN-1032.md
@@ -1,0 +1,3 @@
+- Check for updates now waits while an update or a mute is running, and each
+  of those waits while a check is out, so a check's answer can no longer put
+  rows back as they were before the change.
