+++ ui/src/stores/updates-standing.ts
@@ -10,7 +10,6 @@ import {
 import { type ReadState, readOf, readOrder } from "@/lib/read-state";
 import { settled } from "@/lib/settled";
 import { type PendingFollow, withPending } from "./updates-follow";
-import { commits } from "./updates-writes";
 
 /** The slice a landing writes, and the two it reads. */
 interface Standing {
@@ -61,28 +60,15 @@ export function standingReads(
   };
 
   return {
-    /** Begin an operation that answers with the standing itself, and take
-     *  the landing for that answer.
+    /** Land an answer that reports the standing an operation's own work
+     *  made.
      *
-     *  The answer ranks by when it LANDS rather than when the operation
-     *  began: it reports the state its own work made, so it outranks every
-     *  read still in flight. That claim holds only while nothing else
-     *  commits meanwhile. A command builds its report once, somewhere
-     *  inside itself — `updates_refresh` fetches every mirror and reads the
-     *  standing after — so a commit that landed while it was out may be
-     *  missing from it, and claiming to be newest would put the rows back
-     *  as they were before that commit. Whoever committed read the
-     *  standing back for themselves; theirs is the newer answer.
-     *
-     *  Answers whether it landed, so a caller that lost can read again. */
-    beginOwn: () => {
-      const since = commits.since();
-      return (response: Answer): boolean => {
-        if (commits.stale(since)) return false;
-        land(order.begin(), response);
-        return true;
-      };
-    },
+     *  It ranks by when it LANDS rather than when the operation began: it
+     *  reports the state its own work made, so it outranks every read still
+     *  in flight. That holds because the check and the writes that commit
+     *  exclude each other by flag — nothing can have committed while this
+     *  answer was being built, so there is nothing for it to be missing. */
+    landOwn: (response: Answer) => land(order.begin(), response),
     /** Read the standing again. `reading` is up for as long as one is out,
      *  read off the same ticket that orders the landings, so the page-wide
      *  rule and the ordering rule cannot disagree. */
