+++ ui/src/stores/updates-writes.ts
@@ -1,11 +1,10 @@
-// Every write of the update standing, and the count that says one has
-// happened.
+// Every write of the update standing, and the account each one gives.
 //
-// The count lives here rather than beside each caller because of what it is
-// for: an answer built before a commit must not claim to be newer than it,
-// and the operations that commit are the ones that have to say so. A rule
-// that each new mutation remembers to announce is a rule that gets
-// forgotten — it was, for a whole round, with only one of five paths wired.
+// The account lives here rather than beside each caller because of what it
+// is for: a removal that ran an uninstaller in somebody's repository has to
+// be reported by whatever command took the package away. A rule that each
+// new mutation remembers to announce is a rule that gets forgotten — it
+// was, for a whole round, with one of five paths wired.
 //
 // So the announcement rides on the write itself. A path that goes through
 // this module cannot skip it, and a path that does not go through this
@@ -17,34 +16,13 @@ import {
   type Scope,
   type UpdateRow,
 } from "@/bindings";
-import { invalidations } from "@/lib/read-state";
 import { saying } from "@/lib/undone";
 import type { BulkOutcome } from "@/lib/update-outcome";
 import { type ApplyOutcome, applyRow, applyRows } from "./updates-apply";
 
-/** Writes that have committed. Read by `updates-standing.ts`, which refuses
- *  a landing-time rank to any answer built before the latest one. */
-export const commits = invalidations();
-
-/** Await a write and say a commit may have happened, however it ended.
- *
- *  However it ended, on purpose: the plan writes before it answers, so a
- *  refusal is not proof that nothing landed, and a rejection is proof of
- *  nothing at all. Over-announcing costs an answer its landing-time rank,
- *  which its caller can recover with an ordinary read; under-announcing
- *  lets a stale report overwrite a commit. */
-const committing = async <T>(work: Promise<T>): Promise<T> => {
-  try {
-    // Said on the write for the same reason the commit is announced on it:
-    // a removal that ran an uninstaller in somebody's repository has to be
-    // reported by whatever command took the package away, and a rule each
-    // new mutation remembers is the rule this module already watched get
-    // forgotten. A write whose answer carries no account says nothing.
-    return saying(await work);
-  } finally {
-    commits.moved();
-  }
-};
+/** Await a write and say whatever account its answer carries. */
+const announcing = async <T>(work: Promise<T>): Promise<T> =>
+  saying(await work);
 
 type Report = (error: string) => void;
 
@@ -52,19 +30,19 @@ type Report = (error: string) => void;
 export const writeRow = (
   row: UpdateRow,
   report: Report,
-): Promise<ApplyOutcome> => committing(applyRow(row, report));
+): Promise<ApplyOutcome> => announcing(applyRow(row, report));
 
 /** Bring every place in `rows` current, one apply per place. */
 export const writeRows = (
   rows: UpdateRow[],
   report: Report,
   outcome: BulkOutcome,
-): Promise<void> => committing(applyRows(rows, report, outcome));
+): Promise<void> => announcing(applyRows(rows, report, outcome));
 
 /** Mute or unmute a package's update notices. Answers with the overview it
  *  rebuilt after the write. */
 export const writeIgnored = (row: UpdateRow, ignored: boolean) =>
-  committing(
+  announcing(
     commands.updateSetIgnored(row.scope, row.kind, row.name, row.repo, ignored),
   );
 
@@ -74,11 +52,11 @@ export const writeRev = (
   kind: ItemKind,
   name: string,
   rev: string | null,
-) => committing(commands.packageSetRev(scope, kind, name, rev));
+) => announcing(commands.packageSetRev(scope, kind, name, rev));
 
 /** Bring one package current where it is installed. */
 export const writeUpdate = (scope: Scope, kind: ItemKind, name: string) =>
-  committing(commands.packageUpdate(scope, kind, name));
+  announcing(commands.packageUpdate(scope, kind, name));
 
 /** Keep an edited place's files as a fork of its own. */
 export const writeFork = (
@@ -86,7 +64,7 @@ export const writeFork = (
   kind: ItemKind,
   name: string,
   harness: HarnessId,
-) => committing(commands.packageFork(scope, kind, name, harness));
+) => announcing(commands.packageFork(scope, kind, name, harness));
 
 /** Drop an edited place's edits, moving its hold along where it has one. */
 export const writeDiscardEdits = (
@@ -94,7 +72,7 @@ export const writeDiscardEdits = (
   kind: ItemKind,
   name: string,
   rev: string | null,
-) => committing(commands.applyDiscardEdits(scope, kind, name, rev));
+) => announcing(commands.applyDiscardEdits(scope, kind, name, rev));
 
 /** Keep an edited place's files under a new name, and let the source's
  *  newest version back in under the original. */
@@ -106,4 +84,4 @@ export const writeForkBeside = (
   own: string,
   rev: string | null,
 ) =>
-  committing(commands.packageForkBeside(scope, kind, name, harness, own, rev));
+  announcing(commands.packageForkBeside(scope, kind, name, harness, own, rev));
