+++ ui/src/stores/updates.ts
@@ -35,8 +35,11 @@ interface UpdatesState {
    *  this — the check runs offline on load, so "everything is up to date"
    *  needs the age of the fetch it rests on beside it. */
   lastFetched: number | null;
+  /** True while a write that commits is running. */
   busy: boolean;
-  /** True while a mirror fetch is running — the explicit "check". */
+  /** True while a mirror fetch is running — the explicit "check". It and
+   *  `busy` exclude each other: a fetch builds its report once, so a commit
+   *  landing while it is out would be missing from it. */
   checking: boolean;
   /** True while a read of the standing is on its way. Read off the same
    *  ticket that orders the landings, so the page-wide rule and the
@@ -71,11 +74,11 @@ export const useUpdatesStore = create<UpdatesState>((set, get) => {
 
   const reportUpdate = (error: string) => showError(UPDATE_ERROR_TITLE, error);
 
-  const { beginOwn, reload } = standingReads(set, get);
+  const { landOwn, reload } = standingReads(set, get);
 
-  /** What a commit-applying action says instead of running. Each captures
-   *  values off the rows it was handed, so it may run only against rows a
-   *  read confirmed and outside a scope a follow switch is settling in. */
+  /** What a commit-applying action says instead of running: the rows it
+   *  was handed are not ones to act on, or a check is out whose report
+   *  would land over whatever this committed. */
   const needsCheck = () =>
     showError(UPDATE_ERROR_TITLE, UPDATE_NEEDS_CHECK_NOTE);
 
@@ -93,18 +96,18 @@ export const useUpdatesStore = create<UpdatesState>((set, get) => {
 
     check: async () => {
       // A check already running answers this click too; a second fetch would
-      // cost the network twice for one answer.
-      if (get().checking) return;
+      // cost the network twice for one answer. A write already running is
+      // the other half: the fetch builds its report once, so a commit that
+      // lands while it is out would not be in it, and landing it would put
+      // the rows back as they were before that commit.
+      if (get().checking || get().busy) return;
       set({ checking: true });
-      const landOwn = beginOwn();
       try {
         const response = await settled(commands.updatesRefresh());
         // The fetch reads the standing after fetching every mirror, so it
-        // ranks by when it lands — unless something committed while it was
-        // out, which its report would not carry. The mirrors are fetched
-        // either way, so an ordinary read picks them up ranked behind
-        // whatever committed.
-        if (!landOwn(response)) await reload();
+        // ranks by when it lands — and nothing could have committed behind
+        // it, because a write out is what would have refused this check.
+        landOwn(response);
         if (response.status === "error")
           showError(UPDATE_ERROR_TITLE, response.error);
       } finally {
@@ -184,17 +187,24 @@ export const useUpdatesStore = create<UpdatesState>((set, get) => {
     }),
 
     setIgnored: async (row, ignored) => {
-      const response = await settled(writeIgnored(row, ignored));
-      if (response.status === "error")
-        showError(UPDATE_ERROR_TITLE, response.error);
-      // The command answers with the overview it rebuilt, and this reads
-      // it again anyway. That report carries this write but not one that
-      // landed beside it, and a single count cannot tell an operation's
-      // own commit from another's — so rather than reason about whose bump
-      // was whose, the ordinary read takes it, ranked behind whatever
-      // committed. The command can also persist the preference and then
-      // fail building the overview, which the same read answers.
-      await reload();
+      // The mute captures nothing off the row, so `rowUnsettled` is not what
+      // bars it. What bars it is the check: a report built before this
+      // commit must not land after it, and refusing here is what keeps the
+      // fetch from being out at all. `busy` is up for the same reason in
+      // reverse — while this write is out, a check refuses.
+      if (get().checking || get().busy) return needsCheck();
+      set({ busy: true });
+      try {
+        const response = await settled(writeIgnored(row, ignored));
+        if (response.status === "error")
+          showError(UPDATE_ERROR_TITLE, response.error);
+        // The command answers with the overview it rebuilt, and this reads
+        // it again anyway: the command can persist the preference and then
+        // fail building the overview, which only the read answers.
+        await reload();
+      } finally {
+        set({ busy: false });
+      }
     },
   };
 });
